requirements.txt
streamlit>=1.32.0
plotly>=5.19.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0
yfinance>=0.2.37
entsoe-py>=0.6.0
scikit-learn>=1.4.0
statsmodels>=0.14.0
feedparser>=6.0.0
transformers>=4.38.0
torch>=2.2.0
